fix: resolve build errors and upgrade to compileSdk 36
- MeshRepository: rename private deliveryStatsStore to _deliveryStatsStore
  to avoid conflicting declarations with the public property
- DeviceListFragment: fix binding reference toolbarDevices → toolbar
  (matches @+id/toolbar in fragment_device_list.xml)
- RssiHistoryFragment: fix type mismatch List<Float> → List<Int> for
  setReadings() (RssiSparklineView accepts List<Int>)
- app/build.gradle.kts: bump compileSdk/targetSdk 34 → 36 (only SDK 36
  platform is installed on this machine)
- gradle/wrapper/gradle-wrapper.jar: replace empty jar committed by Copilot
  with a valid one (Gradle 8.6 wrapper)

Verified: BUILD SUCCESSFUL, produces app-debug.apk (7.1 MB)

+++ app/src/main/java/com/turbomesh/computingmachine/ui/history/RssiHistoryFragment.kt
@@ -51,7 +51,7 @@ class RssiHistoryFragment : Fragment() {
             viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
                 launch {
                     viewModel.rssiHistory.collect { entries ->
-                        val readings = entries.map { it.rssi.toFloat() }
+                        val readings = entries.map { it.rssi }
                         binding.rssiSparklineHistory.setReadings(readings)
                         adapter.submitList(entries)
                     }
